Added pp+star calculation and embed creation

+++ src/database/models/ratios.rs
@@ -2,7 +2,7 @@ use sqlx::FromRow;
 
 #[derive(Debug, FromRow)]
 pub struct Ratios {
-    pub scores: Vec<i16>,
+    pub scores: Vec<i8>,
     pub ratios: Vec<f32>,
     pub misses: Vec<f32>,
 }
